refactor!: simplifies and extends typings for Event types
BREAKING CHANGE: reworks the API and types for how Events are used
within this library. Removes the existing `EventType` in favor of a
simpler version where event properties are expressed directly on the
enum variants.

Fixes some composition issues specifically with message events. Adds
some additional types to better distinguish between basic messages and
message subtypes that are used in the Slack Events API.

Introduces the `assert-json-diff` crate to validate serialization and
deserialization of different types